Module bgparse::bmp

BGP Monitoring Protocol

Structs

Initiation

MessageIter

PeerDownNotification

PeerUpNotification

PerPeer

RouteMirroring

RouteMonitoring

RouterInfoIter

StatisticsIter

StatisticsReport

Termination

UnknownStatistic

Enums

Bmp

RouterInfo

Statistic

Constants

BMP_FLAG_IPV6
The V flag indicates the the Peer address is an IPv6 address. For IPv4 peers this is set to 0.

BMP_FLAG_L
The L flag, if set to 1, indicates the message reflects the post-policy Adj-RIB-In (i.e., its path attributes reflect the application of inbound policy). It is set to 0 if the message reflects the pre-policy Adj-RIB-In. Locally-sourced routes also carry an L flag of 1. See Section 5 for further detail. This flag has no significance when used with route mirroring messages (Section 4.7).

BMP_FLAG_LEGACY_AS
The A flag, if set to 1, indicates the message is formatted using the legacy two-byte AS_PATH format. If set to 0, the message is formatted using the four-byte AS_PATH format [RFC6793]. A BMP speaker MAY choose to propagate the AS_PATH information as received from its peer, or it MAY choose to reformat all AS_PATH information into four-byte format regardless of how it was received from the peer. In the latter case, AS4_PATH or AS4_AGGREGATOR path attributes SHOULD NOT be sent in the BMP UPDATE message. This flag has no significance when used with route mirroring messages (Section 4.7).

BMP_MSG_INIT

BMP_MSG_PEERDOWN

BMP_MSG_PEERUP

BMP_MSG_ROUTEMIRROR

BMP_MSG_ROUTEMON

BMP_MSG_STATREPORT

BMP_MSG_TERM

BMP_PEER_GLOBAL

BMP_PEER_LOCAL

BMP_PEER_RD

Traits

Messages

PeerInfo